Require logged-in user, or viewid, for details.php
The details.php page, if it includes a "viewid" parameter,
shows the cloud stuff from that page. If it includes no
"viewid", then it shows the cloud stuff from the current
user's cloud token.

If a logged-out user access details.php, they have no token,
and this causes an unwieldy error message. So it's better
to just say, if you're a logged-out user, you must provide
a view id.

Change-Id: Id9e59908c8c3db9aad3df987eb32a9fff8753f1c